// src/app/models/GroupRole.ts
export class GroupRole {
    id: number;
    screenId: number;
    componentName?: string;
    urlPath: string;
    iconName: string;
    canShow: boolean;
    groupId: number;
    groupName: string;
    view: boolean;
    add: boolean;
    update: boolean;
    delete: boolean;
    print: boolean;
    isReport: boolean;
    screenGroupsAuthoritiesId: number;
}